+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Throw errors when template keys are missing.
+
 ## [0.2.3] - 2021-01-28
 
 ### Fixed